+++ src/apps/polynomial/polynomial.module.ts
@@ -8,14 +8,13 @@ import { PolynomialContractFactory } from './contracts';
 import { PolynomialApiHelper } from './helpers/polynomial.api';
 import { OptimismPolynomialCallSellingVaultQueueContractPositionFetcher } from './optimism/polynomial.call-selling-vault-queue.contract-position-fetcher';
 import { OptimismPolynomialCallSellingVaultTokenFetcher } from './optimism/polynomial.call-selling-vault.token-fetcher';
+import { OptimismPolynomialPerpContractPositionFetcher } from './optimism/polynomial.perp.contract-position-fetcher';
 import { OptimismPolynomialPutSellingVaultQueueContractPositionFetcher } from './optimism/polynomial.put-selling-vault-queue.contract-position-fetcher';
 import { OptimismPolynomialPutSellingVaultTokenFetcher } from './optimism/polynomial.put-selling-vault.token-fetcher';
-import { OptimismPolynomialPerpContractPositionFetcher } from './optimism/polynomial.perp.contract-position-fetcher';
 
 @Module({
   providers: [
     PolynomialApiHelper,
-
     PolynomialContractFactory,
     SynthetixContractFactory,
     OptimismPolynomialCallSellingVaultTokenFetcher,
@@ -25,4 +24,4 @@ import { OptimismPolynomialPerpContractPositionFetcher } from './optimism/polyno
     OptimismPolynomialPerpContractPositionFetcher,
   ],
 })
-export class PolynomialAppModule extends AbstractApp() { }
+export class PolynomialAppModule extends AbstractApp() {}